pyproject.toml
[project]
name = "photo-fix-ml"
version = "0.1.0"
description = "server"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "fastapi>=0.128.0",
    "onnxruntime>=1.23.2",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.21",
    "rembg>=2.0.69",
    "ultralytics==8.3.23",
    "uvicorn>=0.40.0",
    "werkzeug>=3.1.5",
]
